Refine the ci tests of the process pool
